+++ NEWS.md
@@ -1,5 +1,7 @@
 # themis (development version)
 
+* export `smote()` function to users.
+
 # themis 0.1.3
 
 * Steps that use nearest neighbors gives cleaner errors.